requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
# Comment out psycopg2 if you only need SQLite
psycopg2-binary==2.9.10
SQLAlchemy==2.0.28
gunicorn==21.2.0
Werkzeug==2.2.3
email-validator==2.1.0
passlib==1.7.4
bcrypt==4.0.1
Jinja2==3.1.2
requests==2.31.0
python-dateutil==2.8.2
six==1.16.0
# Testing and development tools (optional for production)
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0